Consolidate forecasting utilities into single script
